third_party/WebKit/LayoutTests/fast/js/kde/string-1-n.html

<!DOCTYPE html>

<script src="../../../resources/js-test.js"></script>
<script>
description("KDE JS Test");
expectError();
</script>
<script src="script-tests/string-1-n.js"></script>
<script>
shouldHaveHadError("SyntaxError");
</script>
