LayoutTests/fast/js/kde/string-1-n-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unrecognized token '"unterminated string'
KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.successfullyParsed is undefined.

TEST COMPLETE
